pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pyclawmem"
version = "0.1.0"
description = "Memory, personality, and context continuity for LLM agents"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34",
    "numpy>=1.26",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
